+++ zcash_client_backend/src/data_api/testing/transparent.rs
@@ -131,6 +131,8 @@ where
         outpoint.clone(),
         txout.clone(),
         Some(height_1),
+        crate::TransferType::Incoming,
+        account_id,
         Some(TransparentKeyScope::EXTERNAL),
     )
     .unwrap();
@@ -163,6 +165,8 @@ where
         outpoint,
         txout,
         Some(height_2),
+        crate::TransferType::Incoming,
+        account_id,
         Some(TransparentKeyScope::EXTERNAL),
     )
     .unwrap();
@@ -254,6 +258,8 @@ where
         OutPoint::fake(),
         txout,
         Some(height),
+        crate::TransferType::Incoming,
+        account.id(),
         Some(TransparentKeyScope::EXTERNAL),
     )
     .unwrap();
@@ -408,6 +414,8 @@ where
             outpoint,
             txout.clone(),
             Some(height),
+            crate::TransferType::Incoming,
+            account.id(),
             Some(TransparentKeyScope::EXTERNAL),
         )
         .unwrap();
@@ -497,6 +505,8 @@ where
         OutPoint::fake(),
         txout,
         Some(height),
+        crate::TransferType::Incoming,
+        account.id(),
         Some(TransparentKeyScope::EXTERNAL),
     )
     .unwrap();
@@ -865,7 +875,15 @@ where
     let value = Zatoshis::const_from_u64(50_000);
     let outpoint = OutPoint::fake();
     let txout = TxOut::new(value, taddr.script().into());
-    let utxo = WalletTransparentOutput::from_parts(outpoint, txout, Some(height), None).unwrap();
+    let utxo = WalletTransparentOutput::from_parts(
+        outpoint,
+        txout,
+        Some(height),
+        crate::TransferType::Incoming,
+        account_id,
+        None,
+    )
+    .unwrap();
     st.wallet_mut()
         .put_received_transparent_utxo(&utxo)
         .unwrap();
@@ -941,8 +959,15 @@ where
     let value = Zatoshis::from_u64(100000).unwrap();
     let height = st.wallet().chain_height().unwrap().unwrap();
     let txout = TxOut::new(value, taddr.script().into());
-    let utxo =
-        WalletTransparentOutput::from_parts(OutPoint::fake(), txout, Some(height), None).unwrap();
+    let utxo = WalletTransparentOutput::from_parts(
+        OutPoint::fake(),
+        txout,
+        Some(height),
+        crate::TransferType::Incoming,
+        account_id,
+        None,
+    )
+    .unwrap();
     st.wallet_mut()
         .put_received_transparent_utxo(&utxo)
         .unwrap();
@@ -1196,7 +1221,15 @@ where
     let value = Zatoshis::const_from_u64(50_000);
     let outpoint = OutPoint::fake();
     let txout = TxOut::new(value, taddr.script().into());
-    let utxo = WalletTransparentOutput::from_parts(outpoint, txout, Some(height), None).unwrap();
+    let utxo = WalletTransparentOutput::from_parts(
+        outpoint,
+        txout,
+        Some(height),
+        crate::TransferType::Incoming,
+        account_id,
+        None,
+    )
+    .unwrap();
     st.wallet_mut()
         .put_received_transparent_utxo(&utxo)
         .unwrap();
@@ -1269,8 +1302,15 @@ where
     let value = Zatoshis::from_u64(100000).unwrap();
     let height = st.wallet().chain_height().unwrap().unwrap();
     let txout = TxOut::new(value, taddr.script().into());
-    let utxo =
-        WalletTransparentOutput::from_parts(OutPoint::fake(), txout, Some(height), None).unwrap();
+    let utxo = WalletTransparentOutput::from_parts(
+        OutPoint::fake(),
+        txout,
+        Some(height),
+        crate::TransferType::Incoming,
+        account_id,
+        None,
+    )
+    .unwrap();
     st.wallet_mut()
         .put_received_transparent_utxo(&utxo)
         .unwrap();